refactor: extract WebClient.Builder and ObjectMapper beans to dedicated config
LoadTesterApplication now only owns the @SpringBootApplication entry point.
Moves the ObjectMapper and WebClient.Builder bean declarations to a new
WebBeansConfig under io.camunda.zeebe.config and trims the rationale to a
single class-level javadoc. Spring's component scan picks up the new
@Configuration without an explicit @Import.
